Move shared classes to common module
